requirements.txt
django==4.2.4
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
django-environ==0.11.2
django-cors-headers==4.3.0
Pillow==10.1.0
psycopg2==2.9.9